feature: converted project to typescript
