--- src/App.js
@@ -1,388 +1,0 @@
-import React, { useState, useEffect } from 'react';
-//import {  ExternalLink } from 'lucide-react';
-import { Menu, X, Github, Linkedin, Mail } from 'lucide-react';
-import { personalInfo, experience, education, skills } from './data/portfolioData';
-import profileImage from './data/profile.jpeg';
-
-const Portfolio = () => {
-  const [isMenuOpen, setIsMenuOpen] = useState(false);
-  const [activeSection, setActiveSection] = useState('home');
-  const [isHomeImageVisible, setIsHomeImageVisible] = useState(true);
-
-  // Handle scroll to update active section and home image visibility
-  useEffect(() => {
-    const handleScroll = () => {
-      const sections = ['home', 'experience', 'education', 'skills', 'projects', 'contact'];
-      const scrollPosition = window.scrollY + 100;
-
-      // Check if home section image is visible
-      const homeSection = document.getElementById('home');
-      if (homeSection) {
-        const homeImage = homeSection.querySelector('#profile-image');
-        if (homeImage) {
-          const imageRect = homeImage.getBoundingClientRect();
-          // Image is visible if it's in the viewport
-          setIsHomeImageVisible(imageRect.top < window.innerHeight && imageRect.bottom > 0);
-        }
-      }
-
-      for (const section of sections) {
-        const element = document.getElementById(section);
-        if (element) {
-          const { offsetTop, offsetHeight } = element;
-          if (scrollPosition >= offsetTop && scrollPosition < offsetTop + offsetHeight) {
-            setActiveSection(section);
-            break;
-          }
-        }
-      }
-    };
-
-    window.addEventListener('scroll', handleScroll);
-    return () => window.removeEventListener('scroll', handleScroll);
-  }, []);
-
-  const scrollToSection = (sectionId) => {
-    const element = document.getElementById(sectionId);
-    if (element) {
-      element.scrollIntoView({ behavior: 'smooth' });
-    }
-    setIsMenuOpen(false);
-  };
-
-  const NavLink = ({ href, children, mobile = false }) => (
-    <button
-      onClick={() => scrollToSection(href)}
-      className={`${
-        mobile
-          ? 'block w-full text-left px-4 py-2 text-lg hover:bg-blue-50 transition-colors'
-          : 'px-4 py-2 rounded-lg transition-colors hover:bg-blue-50'
-      } ${
-        activeSection === href
-          ? 'text-blue-600 font-semibold'
-          : 'text-gray-700 hover:text-blue-600'
-      }`}
-    >
-      {children}
-    </button>
-  );
-
-  const profileInitials = personalInfo.name.split(' ').map(n => n[0]).join('').toUpperCase();
-
-  const shouldShowInitials = !personalInfo.img
-
-  return (
-    <div className="min-h-screen bg-white">
-      {/* Navigation Header */}
-      <nav className="fixed top-0 w-full bg-white/90 backdrop-blur-md shadow-sm z-50">
-        <div className="max-w-6xl mx-auto px-4 sm:px-6 lg:px-8">
-          <div className="flex justify-between items-center py-4">
-            <div>
-              {/* Circular Profile Image - Only show when home image is not visible */}
-              {!isHomeImageVisible && (
-                <div className="flex items-center space-x-3">
-                  <div className="w-9 h-9 bg-gradient-to-br from-blue-500 to-purple-600 rounded-full flex items-center justify-center text-white text-sm font-bold overflow-hidden">
-                    {shouldShowInitials ? (
-                      profileInitials
-                    ) : (
-                      <img
-                        id="profile-image"
-                        src={profileImage}
-                        alt={personalInfo.name}
-                        className="w-full h-full object-cover"
-                        onError={(e) => {
-                          console.error('Navigation image failed to load:', personalInfo.img);
-                          // Fallback to initials if image fails
-                          e.target.style.display = 'none';
-                          e.target.parentElement.innerHTML = profileInitials;
-                        }}
-                      />
-                    )}
-                  </div>
-                  <div id="profile-name" className="text-2xl font-bold text-gray-900">
-                    {personalInfo.name}
-                  </div>
-                </div>
-              )}
-            </div>
-            
-            {/* Desktop Navigation */}
-            <div className="hidden md:flex space-x-1">
-              <NavLink href="home">Home</NavLink>
-              <NavLink href="experience">Experience</NavLink>
-              <NavLink href="education">Education</NavLink>
-              <NavLink href="skills">Skills</NavLink>
-              <NavLink href="projects">Projects</NavLink>
-              <NavLink href="contact">Contact</NavLink>
-            </div>
-
-            {/* Mobile Menu Button */}
-            <button
-              onClick={() => setIsMenuOpen(!isMenuOpen)}
-              className="md:hidden p-2 rounded-lg hover:bg-gray-100"
-            >
-              {isMenuOpen ? <X size={24} /> : <Menu size={24} />}
-            </button>
-          </div>
-
-          {/* Mobile Navigation */}
-          {isMenuOpen && (
-            <div className="md:hidden border-t border-gray-200">
-              <div className="py-2">
-                <NavLink href="home" mobile>Home</NavLink>
-                <NavLink href="about" mobile>About</NavLink>
-                <NavLink href="experience" mobile>Experience</NavLink>
-                <NavLink href="education" mobile>Education</NavLink>
-                <NavLink href="skills" mobile>Skills</NavLink>
-                <NavLink href="projects" mobile>Projects</NavLink>
-                <NavLink href="contact" mobile>Contact</NavLink>
-              </div>
-            </div>
-          )}
-        </div>
-      </nav>
-
-      {/* Home Section */}
-      <section id="home" className="pt-16 pb-8 bg-gradient-to-br from-blue-50 to-indigo-100">
-        <div className="max-w-6xl mx-auto px-4 sm:px-6 lg:px-8">
-          <div className="text-center py-10">
-            <div className="w-32 h-32 mx-auto mb-8 bg-gradient-to-br from-blue-500 to-purple-600 rounded-full flex items-center justify-center text-white text-4xl font-bold">
-              {shouldShowInitials ? (
-                profileInitials
-              ) : (
-                <img
-                  src={profileImage}
-                  id="profile-image"
-                  alt={profileInitials}
-                  className="w-full h-full object-cover rounded-full"
-                  onError={(e) => {
-                    console.error('Image failed to load:', personalInfo.img);
-                    e.target.style.display = 'none';
-                    e.target.parentElement.innerHTML = profileInitials;
-                  }}
-                />
-              )}
-            </div>
-            <h1 className="text-5xl sm:text-6xl font-bold text-gray-900 mb-6">
-              {personalInfo.name}
-            </h1>
-            <p className="text-xl sm:text-2xl text-gray-600 mb-8">
-              {personalInfo.title}
-            </p>
-            <div className="text-lg text-gray-800 mb-12 max-w-2xxl mx-auto w-4/5">
-              <p className="text-lg text-gray-600 mb-6">
-              {personalInfo.aboutMe.part1}
-              </p>
-              <p className="text-lg text-gray-600 mb-6">
-              {personalInfo.aboutMe.part2}
-              </p>
-            </div>
-            <div className="inline-flex space-x-4">
-                <a href={personalInfo.social.github} className="text-gray-600 hover:text-blue-600 transition-colors">
-                  <Github size={24} />
-                </a>
-                <a href={personalInfo.social.linkedin} className="text-gray-600 hover:text-blue-600 transition-colors">
-                  <Linkedin size={24} />
-                </a>
-                <a href={personalInfo.social.email} className="text-gray-600 hover:text-blue-600 transition-colors">
-                  <Mail size={24} />
-                </a>
-              </div>
-          </div>
-        </div>
-      </section>
-
-      {/* Experience Section */}
-      <section id="experience" className="py-20 bg-gray-50">
-        <div className="max-w-6xl mx-auto px-4 sm:px-6 lg:px-8">
-          <h2 className="text-4xl font-bold text-center text-gray-900 mb-16">Work Experience</h2>
-          <div className="space-y-12">
-            {experience.map((job, index) => (
-              <div key={index} className="bg-white p-8 rounded-lg shadow-md">
-                <div className="flex flex-col md:flex-row md:justify-between md:items-start mb-4">
-                  <div>
-                    <h3 className="text-2xl font-bold text-gray-900">{job.title}</h3>
-                    <a href={job.link} target="_blank" rel="noreferrer"><p className="text-xl text-blue-600 font-semibold">{job.company}</p></a>
-                  </div>
-                  <span className="text-gray-500 font-medium mt-2 md:mt-0">{job.period}</span>
-                </div>
-                <p className="text-gray-600 mb-4">{job.description}</p>
-                <div className="flex flex-wrap gap-2">
-                  {job.technologies.map((tech, techIndex) => (
-                    <span key={techIndex} className="bg-blue-100 text-blue-800 px-3 py-1 rounded-full text-sm">
-                      {tech}
-                    </span>
-                  ))}
-                </div>
-              </div>
-            ))}
-          </div>
-        </div>
-      </section>
-
-      {/* Education Section */}
-      <section id="education" className="py-20">
-        <div className="max-w-6xl mx-auto px-4 sm:px-6 lg:px-8">
-          <h2 className="text-4xl font-bold text-center text-gray-900 mb-16">Education</h2>
-          <div className="space-y-8">
-            {education.map((edu, index) => (
-              <div key={index} className="bg-white p-8 rounded-lg shadow-md">
-                <div className="flex flex-col md:flex-row md:justify-between md:items-start mb-4">
-                  <div>
-                    <h3 className="text-2xl font-bold text-gray-900">{edu.degree}</h3>
-                    <a href={edu.link} target="_blank" rel="noreferrer"><p className="text-xl text-blue-600 font-semibold">{edu.school}</p></a>
-                  </div>
-                  <span className="text-gray-500 font-medium mt-2 md:mt-0">{edu.period}</span>
-                </div>
-                <p className="text-gray-600">{edu.description}</p>
-              </div>
-            ))}
-          </div>
-        </div>
-      </section>
-
-      {/* Skills Section */}
-      <section id="skills" className="py-20 bg-gray-50">
-        <div className="max-w-6xl mx-auto px-4 sm:px-6 lg:px-8">
-          <h2 className="text-4xl font-bold text-center text-gray-900 mb-16">Skills</h2>
-          <div className="grid md:grid-cols-3 gap-8">
-            {skills.map((skillGroup, index) => (
-              <div key={index} className="bg-white p-8 rounded-lg shadow-md">
-                <h3 className="text-2xl font-bold text-gray-900 mb-6">{skillGroup.category}</h3>
-                <div className="space-y-3">
-                  {skillGroup.skills.map((skill, skillIndex) => (
-                    <div key={skillIndex} className="flex items-center justify-between">
-                      <span className="text-gray-700">{skill}</span>
-                      <div className="w-24 bg-gray-200 rounded-full h-2">
-                        <div 
-                          className="bg-blue-600 h-2 rounded-full" 
-                          style={{ width: `${Math.floor(Math.random() * 30) + 70}%` }}
-                        ></div>
-                      </div>
-                    </div>
-                  ))}
-                </div>
-              </div>
-            ))}
-          </div>
-        </div>
-      </section>
-
-      {/* Projects Section */}
-      {/* <section id="projects" className="py-20">
-        <div className="max-w-6xl mx-auto px-4 sm:px-6 lg:px-8">
-          <h2 className="text-4xl font-bold text-center text-gray-900 mb-16">Featured Projects</h2>
-          <div className="grid md:grid-cols-2 lg:grid-cols-3 gap-8">
-            {projects.map((project, index) => (
-              <div key={index} className="bg-white rounded-lg shadow-md overflow-hidden hover:shadow-lg transition-shadow">
-                <div className={`h-48 ${project.image}`}></div>
-                <div className="p-6">
-                  <h3 className="text-xl font-bold text-gray-900 mb-3">{project.title}</h3>
-                  <p className="text-gray-600 mb-4">{project.description}</p>
-                  <div className="flex flex-wrap gap-2 mb-4">
-                    {project.technologies.map((tech, techIndex) => (
-                      <span key={techIndex} className="bg-gray-100 text-gray-700 px-2 py-1 rounded text-sm">
-                        {tech}
-                      </span>
-                    ))}
-                  </div>
-                  <div className="flex space-x-4">
-                    <a href={project.github} className="text-blue-600 hover:text-blue-800 flex items-center space-x-1">
-                      <Github size={16} />
-                      <span>Code</span>
-                    </a>
-                    <a href={project.demo} className="text-blue-600 hover:text-blue-800 flex items-center space-x-1">
-                      <ExternalLink size={16} />
-                      <span>Live Demo</span>
-                    </a>
-                  </div>
-                </div>
-              </div>
-            ))}
-          </div>
-        </div>
-      </section> */}
-
-      {/* Contact Section */}
-      <section id="contact" className="py-20 bg-gray-50">
-        <div className="max-w-6xl mx-auto px-4 sm:px-6 lg:px-8">
-          <h2 className="text-4xl font-bold text-center text-gray-900 mb-16">Get In Touch</h2>
-          <div className="max-w-2xl mx-auto">
-            <div className="text-center mb-12">
-              <p className="text-lg text-gray-600">
-                I'm open to discussing new opportunities, interesting projects, or just having a chat about technology.
-              </p>
-            </div>
-            <div className="bg-white p-8 rounded-lg shadow-md">
-              <div className="space-y-6">
-                <div className="grid md:grid-cols-2 gap-6">
-                  <div>
-                    <div className="block text-sm font-medium text-gray-700 mb-2">Name</div>
-                    <input
-                      type="text"
-                      className="w-full px-4 py-3 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500 focus:border-transparent"
-                      placeholder="Your Name"
-                    />
-                  </div>
-                  <div>
-                    <div className="block text-sm font-medium text-gray-700 mb-2">Email</div>
-                    <input
-                      type="email"
-                      className="w-full px-4 py-3 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500 focus:border-transparent"
-                      placeholder="your.email@example.com"
-                    />
-                  </div>
-                </div>
-                <div>
-                  <div className="block text-sm font-medium text-gray-700 mb-2">Subject</div>
-                  <input
-                    type="text"
-                    className="w-full px-4 py-3 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500 focus:border-transparent"
-                    placeholder="What's this about?"
-                  />
-                </div>
-                <div>
-                  <div className="block text-sm font-medium text-gray-700 mb-2">Message</div>
-                  <textarea
-                    rows="6"
-                    className="w-full px-4 py-3 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500 focus:border-transparent"
-                    placeholder="Tell me about your project or just say hi!"
-                  ></textarea>
-                </div>
-                <button
-                  className="w-full bg-blue-600 text-white py-3 px-6 rounded-lg hover:bg-blue-700 transition-colors font-medium"
-                >
-                  Send Message
-                </button>
-              </div>
-            </div>
-          </div>
-        </div>
-      </section>
-
-      {/* Footer */}
-      <footer className="bg-gray-900 text-white py-12">
-        <div className="max-w-6xl mx-auto px-4 sm:px-6 lg:px-8">
-          <div className="text-center">
-            <p className="text-gray-400 mb-4">
-              © 2024 {personalInfo.name}. All rights reserved.
-            </p>
-            <div className="flex justify-center space-x-6">
-              <a href={personalInfo.social.github} className="text-gray-400 hover:text-white transition-colors">
-                <Github size={20} />
-              </a>
-              <a href={personalInfo.social.linkedin} className="text-gray-400 hover:text-white transition-colors">
-                <Linkedin size={20} />
-              </a>
-              <a href={personalInfo.social.email} className="text-gray-400 hover:text-white transition-colors">
-                <Mail size={20} />
-              </a>
-            </div>
-          </div>
-        </div>
-      </footer>
-    </div>
-  );
-};
-
-export default Portfolio;
