+++ src/components/index.ts
@@ -1,0 +1,8 @@
+export { Navigation } from './Navigation';
+export { NavLink } from './NavLink';
+export { HomeSection } from './HomeSection';
+export { ExperienceSection } from './ExperienceSection';
+export { EducationSection } from './EducationSection';
+export { SkillsSection } from './SkillsSection';
+export { ContactSection } from './ContactSection';
+export { Footer } from './Footer';
